HMRT http://odos.simor.mech.ntua.gr/index.php
Web Application built on the auction of HMRT project is successfully installed on Apache Server over SSL certificates. In the interim, BLE tunnel notifications are ringed in case of an emergency so that through BLuetooth channel users are informed whether simply they are closed to the exit gates or not.